Move ActivityHeader and put save info in header
