--- ui/src/AtlasUI/components/activity/ActivityHeader/index.js
@@ -1,1 +1,0 @@
-export { ActivityHeader } from './ActivityHeader';
