+++ ui/src/AtlasUI/components/index.js
@@ -1,6 +1,5 @@
 export { Header } from './app/Header';
 export { ActivitySearch } from './activity/ActivitySearch';
-export { ActivityHeader } from './activity/ActivityHeader';
 export { ActivityDetails } from './activity/ActivityDetails';
 export { Print } from './print/Print';
 export { PrintActions } from './print/PrintActions';
