+++ ui/src/KaptureApp/components/activity/ActivityHeader/connect.js
@@ -1,12 +1,15 @@
 import { connect } from 'react-redux';
 
-import { ActivityHeader } from 'AtlasUI/components';
+import { ActivityHeader } from './ActivityHeader';
 import { activity } from 'KaptureApp/store';
 
 const mapState = (state, props) => {
   return {
-    name: activity.selectName(state),
+    lastSaveTime: activity.selectLastSaveTime(state),
     linkUrl: `/activities/${activity.selectId(state)}`,
+    name: activity.selectName(state),
+    saveError: activity.selectSaveError(state),
+    savePending: activity.selectSavePending(state),
   };
 };
