+++ ui/src/KaptureApp/components/activity/ActivityHeader/index.js
@@ -1,0 +1,1 @@
+export { ActivityHeader } from './connect';
