+++ ui/src/KaptureApp/components/index.js
@@ -1,5 +1,5 @@
 export { ActivitySearch } from './ActivitySearchConnected';
-export { ActivityHeader } from './ActivityHeaderConnected';
+export { ActivityHeader } from './activity/ActivityHeader';
 export { ActivityDetails } from './ActivityDetailsConnected';
 export { Editor } from './EditorConnected';
 export { GridTabs } from './GridTabsConnected';
